requirements.txt
alpaca-trade-api>=3.1.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pandas>=1.5.0
requests>=2.28.0
yfinance>=0.2.0
